services/menu/requirements.txt
django==5.2.5
djangorestframework==3.16.1
psycopg2-binary==2.9.10
django-cors-headers==4.7.0
requests==2.31.0
pytest==7.4.2
pytest-django==4.5.2

